+++ pyldapsearch/__main__.py
@@ -10,6 +10,7 @@ from impacket import version
 from pyldapsearch import __version__
 from binascii import unhexlify
 from ldap3 import ANONYMOUS
+from enum import Enum
 import base64
 import logging
 import time
@@ -18,6 +19,11 @@ import json
 import ssl
 import os
 import typer
+
+class SearchScope(str, Enum):
+    BASE = 'BASE'
+    LEVEL = 'LEVEL'
+    SUBTREE = 'SUBTREE'
 
 
 def get_dn(domain):
@@ -264,7 +270,7 @@ class Ldapsearch:
     _separator = '--------------------'
     # bofhound expects some attributes in a certain format
     _base64_attributes = ['nTSecurityDescriptor', 'msDS-GenerationId', 'auditingPolicy', 'dSASignature', 'mS-DS-CreatorSID',
-        'logonHours', 'schemaIDGUID']
+        'logonHours', 'schemaIDGUID', 'cACertificate', 'pKIExpirationPeriod', 'pKIOverlapPeriod']
     _raw_attributes = ['whenCreated', 'whenChanged', 'dSCorePropagationData', 'accountExpires', 'badPasswordTime', 'pwdLastSet',
         'lastLogonTimestamp', 'lastLogon', 'lastLogoff', 'maxPwdAge', 'minPwdAge', 'creationTime', 'lockOutObservationWindow',
         'lockoutDuration']
@@ -272,15 +278,17 @@ class Ldapsearch:
     _ignore_attributes = ['userCertificate']
 
 
-    def __init__(self, ldap_server, ldap_session, query_string, attributes, result_count, search_base, no_query_sd, logs_dir, silent):
+    def __init__(self, ldap_server, ldap_session, scope, query_string, attributes, result_count, search_base, no_query_sd, logs_dir, silent, output):
         self.ldap_server = ldap_server
         self.ldap_session = ldap_session
+        self.scope = scope
         self.query_string = query_string
         self.result_count = result_count
         self.search_base = search_base
         self.no_query_sd = no_query_sd
         self.logs_dir = logs_dir
         self.silent = silent
+        self.output = output
 
         logging.info(f'Distinguished name: {self.search_base}')
         logging.info(f'Filter: {self.query_string}')
@@ -303,6 +311,11 @@ class Ldapsearch:
 
 
     def _printlog(self, line, log=False):
+        # write to -output file if specified
+        if self.output is not None:
+            with open(self.output, 'a') as o :
+                o.write(f'{line}\n')
+
         with open(self.filename, 'a') as f:
             f.write(f'{line}\n')
         if log:
@@ -316,9 +329,9 @@ class Ldapsearch:
         try:
             if 'ntsecuritydescriptor' in self.attributes:
                 controls = ldap3.protocol.microsoft.security_descriptor_control(sdflags=0x07)
-                self.ldap_session.extend.standard.paged_search(self.search_base, self.query_string, attributes=self.attributes, size_limit=self.result_count, controls=controls, paged_size=500, generator=False)
+                self.ldap_session.extend.standard.paged_search(self.search_base, self.query_string, search_scope=self.scope, attributes=self.attributes, size_limit=self.result_count, controls=controls, paged_size=500, generator=False)
             else:
-                self.ldap_session.extend.standard.paged_search(self.search_base, self.query_string, attributes=self.attributes, size_limit=self.result_count, paged_size=500, generator=False)
+                self.ldap_session.extend.standard.paged_search(self.search_base, self.query_string, search_scope=self.scope, attributes=self.attributes, size_limit=self.result_count, paged_size=500, generator=False)
         except (ldap3.core.exceptions.LDAPAttributeError, ldap3.core.exceptions.LDAPInvalidFilterError) as e:
             print()
             logging.critical(f'Error: {str(e)}')
@@ -387,6 +400,7 @@ def main(
     result_count: int = typer.Option(0, '-limit', help='Limit the number of results to return', rich_help_panel='Search Options'),
     domain_controller: str = typer.Option('', '-dc-ip', help='Domain controller IP or hostname to query', rich_help_panel='Connection Options'),
     distinguished_name: str = typer.Option('', '-base-dn', help='Search base distinguished name to use. Default is base domain level', rich_help_panel='Search Options'),
+    scope: SearchScope = typer.Option(SearchScope.SUBTREE, '-scope', help='Scope the query has to be performed', case_sensitive=False, rich_help_panel='Search Options'),
     no_sd: bool = typer.Option(False, '-no-sd', help='Do not add nTSecurityDescriptor as an attribute queried by default. Reduces console output significantly', rich_help_panel='Search Options'),
     debug: bool = typer.Option(False, '-debug', help='Turn DEBUG output ON'),
     hashes: str = typer.Option(None, '-hashes', metavar="LMHASH:NTHASH", help='NTLM hashes, format is LMHASH:NTHASH', rich_help_panel='Connection Options'),
@@ -402,6 +416,7 @@ def main(
     no_smb: bool = typer.Option(False, '-no-smb', help='Do not make a SMB connection to the DC to get its hostname (useful for -k). '
                                         'Requires a hostname to be provided with -dc-ip',
                                         rich_help_panel='Connection Options'),
+    output : str = typer.Option(None, '-output', help='Result output file for this specific search (logging is still enabled)'),
     silent: bool = typer.Option(False, '-silent', help='Do not print query results to console (results will still be logged)', rich_help_panel='Search Options')):
     '''
     Tool for issuing manual LDAP queries which offers bofhound compatible output
@@ -410,7 +425,7 @@ def main(
     print(version.BANNER)
     logger.init()
 
-    logging.info(f'pyldapsearch v{__version__} - @Tw1sm\n')
+    logging.info(f'pyldapsearch v{__version__} - Tw1sm\n')
 
     domain, username, password = parse_credentials(target)
     
@@ -462,7 +477,7 @@ def main(
         ldap_server, ldap_session = init_ldap_session(domain=domain, username=username, password=password, lmhash=lm_hash, 
                                                         nthash=nt_hash, kerberos=kerberos, domain_controller=domain_controller, 
                                                         ldaps=ldaps, hashes=hashes, aesKey=aesKey, no_smb=no_smb, channel_binding=channel_binding)
-        ldapsearch = Ldapsearch(ldap_server, ldap_session, filter, attributes, result_count, search_base, no_sd, logs_dir, silent)
+        ldapsearch = Ldapsearch(ldap_server, ldap_session, scope.value, filter, attributes, result_count, search_base, no_sd, logs_dir, silent, output)
         logging.debug('LDAP bind successful')
     except ldap3.core.exceptions.LDAPSocketOpenError as e: 
         if 'invalid server address' in str(e):